requirements.txt
pandas
numpy
flask
nltk
scikit-learn
streamlit
sentence-transformers 
faiss-cpu
seaborn
matplotlib